Changed endgame script and scence to different spelling to allow to switch to camelcase
