+++ Assembly-CSharp.csproj
@@ -66,7 +66,7 @@
      <Compile Include="Assets\Code\Classes\Roboticon.cs" />
      <Compile Include="Assets\Code\Classes\Timeout.cs" />
      <Compile Include="Assets\Code\Scripts\GUI\canvasScript.cs" />
-     <Compile Include="Assets\Code\Scripts\GUI\endGameScript.cs" />
+     <Compile Include="Assets\Code\Scripts\GUI\enndGameScript.cs" />
      <Compile Include="Assets\Code\Scripts\GUI\helpBoxScript.cs" />
      <Compile Include="Assets\Code\Scripts\GUI\marketScript.cs" />
      <Compile Include="Assets\Code\Scripts\GUI\roboticonGuiElementScript.cs" />
